Hazard Model
We study a comprehensive clinical trial conducted by the Radiation Therapy Oncology Group in the United States. The investigation is to assess the efficacy of these two treatment approaches. In performing the investigation, a Kaplan-Meier estimation was employed extensively to reveal insufficient difference of the treatment. Further analysis of stratification proves helpful in investigating patient outcomes and in the development of a Cox Proportional Hazard model which quantifies factors the general public may already perceive play a pivotal role in predicting outcomes.